+++ src/main/java/roomescape/controller/ReservationController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.service.dto.ReservationResponse;
-import roomescape.service.dto.SaveReservationRequest;
+import roomescape.service.dto.response.ReservationResponse;
+import roomescape.service.dto.request.SaveReservationRequest;
 import roomescape.service.reservation.ReservationCreateService;
 import roomescape.service.reservation.ReservationDeleteService;
 import roomescape.service.reservation.ReservationFindService;
