+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.ReservationTime;
-import roomescape.service.dto.ReservationTimeIsBookedResponse;
-import roomescape.service.dto.ReservationTimeResponse;
-import roomescape.service.dto.SaveReservationTimeRequest;
+import roomescape.service.dto.response.ReservationTimeIsBookedResponse;
+import roomescape.service.dto.response.ReservationTimeResponse;
+import roomescape.service.dto.request.SaveReservationTimeRequest;
 import roomescape.service.reservationtime.ReservationTimeCreateService;
 import roomescape.service.reservationtime.ReservationTimeDeleteService;
 import roomescape.service.reservationtime.ReservationTimeFindService;
