+++ src/main/java/roomescape/controller/ThemeController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Theme;
-import roomescape.service.dto.SaveThemeRequest;
-import roomescape.service.dto.ThemeResponse;
+import roomescape.service.dto.request.SaveThemeRequest;
+import roomescape.service.dto.response.ThemeResponse;
 import roomescape.service.theme.ThemeCreateService;
 import roomescape.service.theme.ThemeDeleteService;
 import roomescape.service.theme.ThemeFindService;
