+++ src/main/java/roomescape/service/reservation/ReservationCreateService.java
@@ -7,7 +7,7 @@ import roomescape.domain.Theme;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.dto.SaveReservationRequest;
+import roomescape.service.dto.request.SaveReservationRequest;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
