+++ src/test/java/roomescape/service/dto/SaveReservationRequestTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.service.dto.request.SaveReservationRequest;
 
 import java.time.LocalDate;
